Update Cargo.lock to change yral-common source URLs to use .git extension; refactor notification component to improve structure and remove unused code.
